Security Configuration added. CORS handling moved to SecurityConfiguration class. Authentication provided for UserRestController and PackageRestController.

+++ ajdu_restful_api/src/main/java/ajdu_restful_api/controller/PackageRestController.java
@@ -7,7 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -22,8 +22,7 @@ import ajdu_restful_api.service.ServiceService;
 import ajdu_restful_api.service.UserService;
 
 @RestController
-@CrossOrigin(origins={"http://localhost:9000", "http://localhost:4200", "http://localhost:8080"}, maxAge=4800)
-public class PackageRestController {
+public class PackageRestController extends AuthenticatedRestController {
 
 	@Autowired
 	PackageService packService;
@@ -39,34 +38,47 @@ public class PackageRestController {
 	}
 	
 	@RequestMapping(value="/packages",method=RequestMethod.POST,produces=MediaType.APPLICATION_JSON_UTF8_VALUE)
-	public ResponseEntity<Package> createPackage(@RequestBody Package pack) {
-		if(pack.getUser() == null || pack.getUser().getId() == null) {
-			return new ResponseEntity<Package>(HttpStatus.BAD_REQUEST);
+	public ResponseEntity<Package> createPackage(@RequestBody Package pack, Authentication auth) {
+		if(pack.getUser() == null || pack.getUser().getId() == null)
+			return new ResponseEntity<Package>(HttpStatus.BAD_REQUEST);		
+		
+		User u = userService.findUser(pack.getUser().getId());
+		
+		if(!hasPermission(auth, u.getLogin()))
+			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+		
+		if(u.getPack() != null)
+			return new ResponseEntity<Package>(HttpStatus.CONFLICT);
+		else {			
+			packService.save(pack);
+			return new ResponseEntity<Package>(pack, HttpStatus.CREATED);
 		}
-		else {
-			User u = userService.findUser(pack.getUser().getId());
-			if(u.getPack() != null)
-				return new ResponseEntity<Package>(HttpStatus.CONFLICT);
-			else {			
-				packService.save(pack);
-				return new ResponseEntity<Package>(pack, HttpStatus.CREATED);
-			}
-		}		
+			
 	}
 	
 	
 	@RequestMapping(value="/packages/{id}",method=RequestMethod.GET,produces=MediaType.APPLICATION_JSON_UTF8_VALUE)
-	public ResponseEntity<Package> getPackage(@PathVariable int id) {
+	public ResponseEntity<Package> getPackage(@PathVariable int id, Authentication auth) {
 		Package p = packService.findPackage(id);
-		if(p != null)
-			return new ResponseEntity<Package>(p,HttpStatus.OK);
+		if(p != null){
+			User u = userService.findUser(p.getUser().getId());
+			if(!hasPermission(auth, u.getLogin()))
+				return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+			else return new ResponseEntity<Package>(p,HttpStatus.OK); 
+		}
 		else return new ResponseEntity<Package>(HttpStatus.NOT_FOUND);
 	}
 	
 	@RequestMapping(value="/packages/{id}",method=RequestMethod.PUT,consumes=MediaType.APPLICATION_JSON_UTF8_VALUE )
-	public ResponseEntity<Package> updatePackage(@PathVariable int id, @RequestBody Package pack) {
-		Package p = packService.findPackage(id);
+	public ResponseEntity<Package> updatePackage(@PathVariable int id, @RequestBody Package pack, Authentication auth) {
+		
+		Package p = packService.findPackage(id);		
+		
 		if(p != null) {
+			User u = userService.findUser(p.getUser().getId());
+			if(!hasPermission(auth, u.getLogin()))
+				return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+			
 			List<Service> services = new ArrayList<Service>();
 			if(pack.getServices() != null) {
 				for(Service serv: pack.getServices()) {
@@ -85,9 +97,13 @@ public class PackageRestController {
 	
 	
 	@RequestMapping(value="/packages/{id}",method=RequestMethod.DELETE)
-	public ResponseEntity<Package> deletePackage(@PathVariable int id) {
+	public ResponseEntity<Package> deletePackage(@PathVariable int id, Authentication auth) {
 		Package p = packService.findPackage(id);
 		if(p != null) {
+			User u = userService.findUser(p.getUser().getId());
+			if(!hasPermission(auth, u.getLogin()))
+				return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+			
 			packService.delete(id);
 			return new ResponseEntity<Package>(HttpStatus.NO_CONTENT);
 		}
@@ -96,22 +112,31 @@ public class PackageRestController {
 	
 	
 	@RequestMapping(value="/packages/{id}/services",method=RequestMethod.GET,produces=MediaType.APPLICATION_JSON_UTF8_VALUE)
-	public ResponseEntity<List<Service>> findAllServicesByPackage(@PathVariable int id){
-		if(packService.findPackage(id) != null) 
-			return new ResponseEntity<List<Service>>(serviceService.findAllServicesByPackage(id), HttpStatus.OK);
+	public ResponseEntity<List<Service>> findAllServicesByPackage(@PathVariable int id, Authentication auth){
+		Package p = packService.findPackage(id);
+		if(p != null) {
+			User u = userService.findUser(p.getUser().getId());
+			if(!hasPermission(auth, u.getLogin()))
+				return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+			else return new ResponseEntity<List<Service>>(serviceService.findAllServicesByPackage(id), HttpStatus.OK);
+		}
 		else 
 			return new ResponseEntity<List<Service>>(HttpStatus.NOT_FOUND);
 	}
 	
 	
 	@RequestMapping(value="/packages/{id}/services", method=RequestMethod.POST,consumes=MediaType.APPLICATION_JSON_UTF8_VALUE)
-	public ResponseEntity<List<Service>> addServiceToPackage(@RequestBody Service service, @PathVariable int id) {
+	public ResponseEntity<List<Service>> addServiceToPackage(@RequestBody Service service, @PathVariable int id, Authentication auth) {
 		Package p = packService.findPackage(id);
 		if(p==null)
 			return new ResponseEntity<List<Service>>(HttpStatus.NOT_FOUND);
 		else if(service.getId() == null)
 			return new ResponseEntity<List<Service>>(HttpStatus.BAD_REQUEST);
 		else {
+			User u = userService.findUser(p.getUser().getId());
+			if(!hasPermission(auth, u.getLogin()))
+				return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+			
 			Service s = serviceService.findService(service.getId());
 			if(s == null)
 				return new ResponseEntity<List<Service>>(HttpStatus.NOT_FOUND);
@@ -129,12 +154,16 @@ public class PackageRestController {
 	
 	@RequestMapping(value="/packages/{packageId}/services/{serviceId}",
 			method=RequestMethod.DELETE,consumes=MediaType.APPLICATION_JSON_UTF8_VALUE)
-	public ResponseEntity<Service> deleteServiceFromPackage(@PathVariable int packageId, @PathVariable int serviceId) {
+	public ResponseEntity<Service> deleteServiceFromPackage(@PathVariable int packageId, @PathVariable int serviceId, Authentication auth) {
 		Package p = packService.findPackage(packageId);
 		Service s = serviceService.findService(serviceId);
 		if(p==null || s==null || !p.getServices().contains(s))
 			return new ResponseEntity<Service>(HttpStatus.NOT_FOUND);
 		else {
+			User u = userService.findUser(p.getUser().getId());
+			if(!hasPermission(auth, u.getLogin()))
+				return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+			
 			p.getServices().remove(s);
 			packService.save(p);
 			return new ResponseEntity<Service>(HttpStatus.OK);
